refactor(canvas): split KonvaObject god component
KonvaObject.tsx had grown to 718 lines bundling four component bodies
(LineObject, ImageObject, BarcodeObject dispatch, KonvaObjectInner)
plus duplicated FT-baseline + 15-dot rotation math. Three changes:

- Pure helpers in textPositionTransforms.ts (objectToDisplay /
  displayToObject) with a round-trip property test that locks the
  inverse property across all rotation × positionType combinations.
  Kills the duplicated math at the render path and the drag-end
  handler.
- LineObject and ImageObject move to their own files and share the
  renderer Props contract via konvaObjectProps.ts.
- The dispatcher narrows obj before passing it down, so each
  per-type renderer takes the narrowed LabelObject variant directly
  and the runtime cast disappears.

KonvaObject.tsx is now 339 lines, focused on dispatch + the shape/
text body in KonvaObjectInner.
